+++ nodejs/examples/01_the_switch_and_the_light/light.js
@@ -11,9 +11,6 @@ const NODE_NAME = 'arete-sdk-01-light';
 
 const PADI_LIGHT_PROFILE = 'padi.light';
 
-// Configure pin for output
-let pin = new Gpio(GPIO23, 'out');
-
 // Connect to Arete control plane
 let client = new Client({
   protocol: 'wss:',
@@ -28,16 +25,24 @@ let system = await client.system();
 let node = await system.node(NODE_ID, NODE_NAME, false);
 let context = await node.context(CONTEXT_ID, CONTEXT_NAME);
 
-// Detect initial desired state, plus future changes to desired state, and try to actualize it
+// Read initial actual state of the light, and sync it with Arete
+let pin = new Gpio(GPIO23, 'in', 'both', { debounceTimeout: 10 });
+let state = pin.readSync();
 let consumer = await context.consumer(PADI_LIGHT_PROFILE);
+consumer.put('cState', state ? '1' : '0');
+console.log('Light is initially', state ? 'ON' : 'OFF');
+
+// Detect initial desired state, plus future changes to desired state, and try to actualize it
 consumer.watch((event) => {
   if (event.property == 'sOut') {
     const desiredState = event.value == '1';
+    let pin = new Gpio(GPIO23, 'out');
     pin.writeSync(desiredState ? Gpio.HIGH : Gpio.LOW, (err) => {
       if (err) {
         throw err;
       }
     });
+    consumer.put('cState', desiredState ? '1' : '0');
     console.log('Light is now', desiredState ? 'ON' : 'OFF');
   }
 });