+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - [#90](https://github.com/project-arete/sdk/issues/90) Consumer and Provider classes/structs both require get, put, and watch capabilities
+- [#86](https://github.com/project-arete/sdk/issues/86) Light examples should provide feedback on actual state
 
 ## [0.1.5] - 2025-09-19
 ### Added